+++ crates/ruma-api/tests/ui/07-error-type-attribute.rs
@@ -15,7 +15,7 @@ ruma_api! {
     response: {}
 
     #[derive(Default)]
-    error: ruma_api::error::Void
+    error: ruma_api::error::MatrixError
 }
 
 fn main() {}